requirements.txt
argcomplete==1.9.3
boto3==1.9.115
botocore==1.12.115
certifi==2019.3.9
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
Markdown==3.0.1
MarkupSafe==1.1.1
placebo==0.9.0
psycopg2==2.7.7
PyJWT==1.7.1
python-dateutil==2.7.3
python-slugify==1.2.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
s3transfer==0.2.0
six==1.11.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.5
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.1
